Add support for linux
Basically I moved all of the code related to exporting/importing/signing ctcert.bin/footer.bin into its own class, Signing.java
This is because I needed more space and logic related to detecting operating systems and doing the correct actions, etc which made it too cluttered to remain in Seedplanter.java

A new binary was also included, linux_ctr-dsiwaretool! Full credit goes to @saibotu for compiling this

Only issue I'm running into right now is compiling by running the `package` plugin of maven. It seems to create a jar with corrupted resources in it, however, opening the jar with 7zip and deleting the resources and putting in the uncorrupted ones seems to work fine so I'm going to push it for now, but I do need to look into this