Rewrite ctr-dsiwaretool (WIP!)
Been coding from 4 to 9 with no break, I think this is enough for today.

So far what I've managed to do is write some helper functions (to get the ECParameterSpec, and to retrieve the ECPoint from a ASN.1 encoded signature). Since these were working, I then tried using them to sign the hashes, but unfortunately it did not work ("Failed to copy data"...).
I believe this is because I am not correctly padding the zeroes in my R and S points. Another reason could also be because I am not signing the APCert "APXXXXXXXXXXXXXXXX" properly (apparently ctr-dsiwaretool signs this as well)

Progress is being made though!

Oh, and I also removed all Windows specific files, since we won't be needing them with this rewrite ;-)